// speeditApp/src/types/typed.ts
// 여러가지 타입이나 인터페이스 선언 ( 한 화면에서 쓸 것들은 여기 선언하지 않아도 됨, 해당 파일에 직접선언해도됨 )

export interface TempBookInfoProps {
  imageUrl: string;
  url: string;
  title: string;
  author: string;
  publishing: string;
  publisher: string;
}